+++ core/src/main/scala/org/apache/spark/util/collection/ExternalAppendOnlyMap.scala
@@ -91,7 +91,8 @@ class ExternalAppendOnlyMap[K, V, C](
   private var _diskBytesSpilled = 0L
   
   // Use getSizeAsKb (not bytes) to maintain backwards compatibility of on units are provided
-  private val fileBufferSize = sparkConf.getSizeAsKb("spark.shuffle.file.buffer", "32k") * 1024
+  private val fileBufferSize = 
+    sparkConf.getSizeAsKb("spark.shuffle.file.buffer", "32k").toInt * 1024
 
   // Write metrics for current spill
   private var curWriteMetrics: ShuffleWriteMetrics = _
